endpoint: move state info to `state.go`
Signed-off by: Ian Vernon <ian@cilium.io>
